create a separate proposal context
