+++ app/proposal/page.tsx
@@ -4,6 +4,7 @@ import { Container, Stack } from '@mantine/core';
 import { useSearchParams } from 'next/navigation';
 import { Layout } from '@/components/Layout/Layout';
 import { ProposalDetailCard } from '@/components/Proposals/ProposalDetailCard';
+import { ProposalProvider } from '@/contexts/ProposalContext';
 
 export default function ProposalsPage() {
   const params = useSearchParams();
@@ -12,7 +13,9 @@ export default function ProposalsPage() {
     <Layout>
       <Container>
         <Stack gap="15">
-          <ProposalDetailCard proposalNumber={proposalNumber} />
+          <ProposalProvider proposalNumber={proposalNumber}>
+            <ProposalDetailCard />
+          </ProposalProvider>
         </Stack>
       </Container>
     </Layout>
