+++ components/Proposals/MarketCard.tsx
@@ -17,20 +17,17 @@ import {
 import numeral from 'numeral';
 import { IconQuestionMark } from '@tabler/icons-react';
 import { NUMERAL_FORMAT } from '@/lib/constants';
-import { useProposal } from '../../hooks/useProposal';
 import { Token, useTokens } from '../../hooks/useTokens';
 import { useTokenAmount } from '../../hooks/useTokenAmount';
-import { ProposalAccountWithKey } from '../../lib/types';
 import { useTokenMint } from '../../hooks/useTokenMint';
 import { useTransactionSender } from '../../hooks/useTransactionSender';
 import { useAutocrat } from '../../contexts/AutocratContext';
 import { getParsedOrders } from '@/lib/openbook';
+import { useProposal } from '@/contexts/ProposalContext';
 
-export function MarketCard({ proposal: fromProposal }: { proposal: ProposalAccountWithKey }) {
-  const { daoTreasury, fetchMarketsInfo } = useAutocrat();
-  const { proposal, markets, mintTokensTransactions, placeOrderTransactions } = useProposal({
-    fromProposal,
-  });
+export function MarketCard() {
+  const { daoTreasury } = useAutocrat();
+  const { proposal, markets, mintTokensTransactions, placeOrderTransactions, fetchMarketsInfo } = useProposal();
   const { amount: baseBalance } = useTokenAmount(markets?.baseVault.underlyingTokenMint);
   const { amount: quoteBalance } = useTokenAmount(markets?.quoteVault.underlyingTokenMint);
   const { tokens } = useTokens();
