--- hooks/useProposal.ts
@@ -1,275 +1,0 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useConnection, useWallet } from '@solana/wallet-adapter-react';
-import { PublicKey, Transaction } from '@solana/web3.js';
-import {
-  getAssociatedTokenAddressSync,
-  createAssociatedTokenAccountInstruction,
-} from '@solana/spl-token';
-import { notifications } from '@mantine/notifications';
-import { ProposalAccountWithKey } from '@/lib/types';
-import { useAutocrat } from '@/contexts/AutocratContext';
-import { useConditionalVault } from '@/hooks/useConditionalVault';
-import { useOpenbookTwap } from './useOpenbookTwap';
-import { useTransactionSender } from './useTransactionSender';
-
-export function useProposal({
-  fromNumber,
-  fromProposal,
-}: {
-  fromNumber?: number | undefined;
-  fromProposal?: ProposalAccountWithKey;
-}) {
-  const {
-    autocratProgram,
-    dao,
-    daoState,
-    daoTreasury,
-    proposals,
-    allMarketsInfo,
-    allOrders,
-    fetchMarketsInfo,
-    fetchOpenOrders,
-  } = useAutocrat();
-  const { connection } = useConnection();
-  const wallet = useWallet();
-  const sender = useTransactionSender();
-  const { placeOrderTransactions } = useOpenbookTwap();
-  const {
-    program: vaultProgram,
-    mintConditionalTokens,
-    getVaultMint,
-    createConditionalTokensAccounts,
-  } = useConditionalVault();
-  const [loading, setLoading] = useState(false);
-  const [metaDisabled, setMetaDisabled] = useState(false);
-  const [usdcDisabled, setUsdcDisabled] = useState(false);
-  const proposal = useMemo<ProposalAccountWithKey | undefined>(
-    () =>
-      proposals?.filter(
-        (t) =>
-          t.account.number === fromNumber ||
-          t.publicKey.toString() === fromProposal?.publicKey.toString(),
-      )[0],
-    [proposals, fromProposal, fromNumber],
-  );
-  const markets = proposal ? allMarketsInfo[proposal.publicKey.toString()] : undefined;
-  const orders = proposal ? allOrders[proposal.publicKey.toString()] : undefined;
-
-  useEffect(() => {
-    if (!orders && proposal && wallet.publicKey) {
-      fetchOpenOrders(proposal, wallet.publicKey);
-    }
-  }, [orders, markets, fetchOpenOrders]);
-
-  useEffect(() => {
-    if (!markets && proposal) {
-      fetchMarketsInfo(proposal);
-    }
-  }, [markets, fetchMarketsInfo]);
-
-  const createTokenAccountsTransactions = useCallback(
-    async (fromBase?: boolean) => {
-      if (!proposal || !markets) {
-        return;
-      }
-      const createAccounts = await createConditionalTokensAccounts(
-        proposal.account,
-        fromBase ? markets.baseVault : markets.quoteVault,
-        fromBase,
-      );
-      const tx = new Transaction().add(...(createAccounts?.ixs ?? []));
-
-      return [tx];
-    },
-    [proposal, markets],
-  );
-
-  const createTokenAccounts = useCallback(
-    async (fromBase?: boolean) => {
-      const txs = await createTokenAccountsTransactions(fromBase);
-      if (!txs || !proposal || !wallet.publicKey) {
-        return;
-      }
-      let error = false;
-      let metaBalance = null;
-      const metaMint = new PublicKey('METADDFL6wWMWEoKTFJwcThTbUmtarRJZjRpzUvkxhr');
-      const quoteVault = await getVaultMint(proposal.account.quoteVault);
-      const baseVault = await getVaultMint(proposal.account.baseVault);
-      const userBasePass = getAssociatedTokenAddressSync(
-        baseVault.conditionalOnFinalizeTokenMint,
-        wallet.publicKey,
-      );
-      const userQuotePass = getAssociatedTokenAddressSync(
-        quoteVault.conditionalOnFinalizeTokenMint,
-        wallet.publicKey,
-      );
-      const metaTokenAccount = getAssociatedTokenAddressSync(metaMint, wallet.publicKey, false);
-
-      try {
-        metaBalance = await connection.getTokenAccountBalance(metaTokenAccount);
-      } catch (err) {
-        console.log('unable to fetch balance for META token account');
-      }
-      try {
-        if (fromBase) {
-          await connection.getTokenAccountBalance(userBasePass);
-        } else {
-          await connection.getTokenAccountBalance(userQuotePass);
-        }
-      } catch (err) {
-        error = true;
-        console.log("turns out the account doesn't exist we can create it");
-      }
-      if (!error) {
-        notifications.show({
-          title: 'Token Accounts Exist',
-          message: "You won't need to press this button again.",
-          autoClose: 5000,
-        });
-        if (fromBase) {
-          setMetaDisabled(true);
-        } else {
-          setUsdcDisabled(true);
-        }
-      }
-
-      if (error) {
-        if (metaBalance === null) {
-          const tx = new Transaction();
-          tx.add(
-            createAssociatedTokenAccountInstruction(
-              wallet.publicKey, // payer
-              metaTokenAccount, // ata
-              wallet.publicKey, // owner
-              metaMint, // mint
-            ),
-          );
-          txs.unshift(tx);
-        }
-        setLoading(true);
-
-        try {
-          await sender.send(txs);
-        } catch (err) {
-          console.error(err);
-        } finally {
-          setLoading(false);
-        }
-      }
-    },
-    [wallet, connection, sender, createTokenAccountsTransactions],
-  );
-
-  const finalizeProposalTransactions = useCallback(async () => {
-    if (!autocratProgram || !proposal || !dao || !daoState || !vaultProgram) return;
-
-    const tx = await autocratProgram.methods
-      .finalizeProposal()
-      .accounts({
-        proposal: proposal.publicKey,
-        openbookTwapPassMarket: proposal.account.openbookTwapPassMarket,
-        openbookTwapFailMarket: proposal.account.openbookTwapFailMarket,
-        dao,
-        daoTreasury,
-        baseVault: proposal.account.baseVault,
-        quoteVault: proposal.account.quoteVault,
-        vaultProgram: vaultProgram.programId,
-      })
-      .transaction();
-
-    return [tx];
-  }, [autocratProgram, proposal, vaultProgram, dao, daoTreasury]);
-
-  const mintTokensTransactions = useCallback(
-    async (amount: number, fromBase?: boolean) => {
-      if (!proposal || !markets || !wallet.publicKey) {
-        return;
-      }
-
-      const mint = await mintConditionalTokens(
-        amount,
-        proposal.account,
-        fromBase ? markets.baseVault : markets.quoteVault,
-        fromBase,
-      );
-      const tx = new Transaction().add(...(mint?.ixs ?? []));
-
-      return [tx];
-    },
-    [proposal, markets],
-  );
-
-  const mintTokens = useCallback(
-    async (amount: number, fromBase?: boolean) => {
-      const txs = await mintTokensTransactions(amount, fromBase);
-      if (!txs || !proposal) {
-        return;
-      }
-
-      setLoading(true);
-
-      try {
-        await sender.send(txs);
-        await fetchMarketsInfo(proposal);
-      } catch (err) {
-        console.error(err);
-      } finally {
-        setLoading(false);
-      }
-    },
-    [connection, sender, mintTokensTransactions],
-  );
-
-  const placeOrder = useCallback(
-    async (amount: number, price: number, limitOrder?: boolean, ask?: boolean, pass?: boolean) => {
-      if (!proposal || !markets) return;
-      const market = pass
-        ? { publicKey: proposal?.account.openbookPassMarket, account: markets?.pass }
-        : { publicKey: proposal?.account.openbookFailMarket, account: markets?.fail };
-      const placeTxs = await placeOrderTransactions(amount, price, market, limitOrder, ask, pass);
-
-      if (!placeTxs || !wallet.publicKey) {
-        return;
-      }
-
-      try {
-        setLoading(true);
-
-        await sender.send(placeTxs);
-        await fetchMarketsInfo(proposal);
-        await fetchOpenOrders(proposal, wallet.publicKey);
-      } catch (err) {
-        console.error(err);
-      } finally {
-        setLoading(false);
-      }
-    },
-    [
-      wallet,
-      proposal,
-      markets,
-      connection,
-      sender,
-      placeOrderTransactions,
-      fetchMarketsInfo,
-      fetchOpenOrders,
-    ],
-  );
-
-  return {
-    proposal,
-    markets,
-    orders,
-    loading,
-    metaDisabled,
-    usdcDisabled,
-    fetchOpenOrders,
-    createTokenAccounts,
-    createTokenAccountsTransactions,
-    finalizeProposalTransactions,
-    mintTokensTransactions,
-    mintTokens,
-    placeOrderTransactions,
-    placeOrder,
-  };
-}
